RTTI: Fixed convertion error using explicit constructors in Any

+++ geUtility/Include/geRTTIType.h
@@ -1637,7 +1637,7 @@ namespace geEngineSDK {
       RTTIPlainField<DataType, ObjectType>*
       newField = ge_new<RTTIPlainField<DataType, ObjectType>>();
 
-      newField->initSingle(name, uniqueId, getter, setter, flags);
+      newField->initSingle(name, static_cast<uint16>(uniqueId), getter, setter, flags);
       addNewField(newField);
     }
